2.24 GUI settings from inside!

+++ src/main/java/harmonised/pmmo/config/Config.java
@@ -73,7 +73,7 @@ public class Config
         public ConfigHelper.ConfigValueListener<Double> xpDropSpawnDistance;
         public ConfigHelper.ConfigValueListener<Double> xpDropOpacityPerTime;
         public ConfigHelper.ConfigValueListener<Double> xpDropMaxOpacity;
-        public ConfigHelper.ConfigValueListener<Integer> xpDropDecayAge;
+        public ConfigHelper.ConfigValueListener<Double> xpDropDecayAge;
         public ConfigHelper.ConfigValueListener<Boolean> showXpDrops;
         public ConfigHelper.ConfigValueListener<Boolean> stackXpDrops;
         public ConfigHelper.ConfigValueListener<Boolean> xpDropsAttachedToBar;
@@ -428,7 +428,7 @@ public class Config
                 this.xpDropDecayAge = subscriber.subscribe(builder
                         .comment( "At what age do xp drops start to decay?" )
                         .translation( "pmmo.xpDropDecayAge" )
-                        .defineInRange( "xpDropDecayAge", 350, 0, 5000) );
+                        .defineInRange( "xpDropDecayAge", 350D, 0, 5000) );
 
                 this.xpDropsAttachedToBar = subscriber.subscribe(builder
                         .comment( "Should xp drops sync up with the bar being open or closed? HIGHLY RECOMMEND TO KEEP FALSE IF YOU ARE MOVING XP DROP POSITIONS" )
